moved "update review summary event" into SectionWrapper from Assessment
